% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_learner.R
\name{df_learner}
\alias{df_learner}
\title{Estimate heterogeneous treatment effect using Doubly Robust Estimation
(Kennedy 2022)}
\usage{
df_learner(X, Y, W, family = "gaussian")
}
\arguments{
\item{X}{matrix of covariates}

\item{Y}{numeric vector of outcomes}

\item{W}{logical vector of treatment states (0, 1)}

\item{family}{character in ("gaussian", "binomial") to pass to \code{cv.glmnet}}
}
\value{
A list of estimates needed for best linear projections of the
conditional average treatment effect for approximately optimal targeting
The list includes observed outcomes (\code{Y}), treatments (\code{W}),
estimates of E[Y | X = x] (\code{Y.hat}) and E[W | X = x] (\code{W.hat}),
and the localized predictions of the causal forest E[Y_1 - Y_0 | X = x]
(\code{tau.hat})
}
\description{
Estimate heterogeneous treatment effect using Doubly Robust Estimation
(Kennedy 2022)
}
\references{
Kennedy, Edward H. (2022) "Towards optimal doubly robust
estimation of heterogeneous causal effects".
\url{https://arxiv.org/abs/2004.14497}.
}
