% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr_learner.R
\name{dr_learner}
\alias{dr_learner}
\title{Estimate heterogeneous treatment effect using Doubly Robust Estimation
(Kennedy 2022) using \code{cv.glmnet} for estimate construction}
\usage{
dr_learner(X, Y, W, family = "gaussian", ...)
}
\arguments{
\item{X}{matrix of covariates}

\item{Y}{numeric vector of outcomes}

\item{W}{numeric vector of treatment states [0, 1]. If a logical vector is
supplied, will coerce to numeric with \code{FALSE = 0} and \code{TRUE = 1}.}

\item{family}{character in ("gaussian", "binomial") to pass to \code{cv.glmnet}}

\item{...}{arguments to pass to \code{cv.glmnet}}
}
\value{
A list of estimates needed for best linear projections of the
conditional average treatment effect for approximately optimal targeting
The list includes observed outcomes (\code{Y}), treatments (\code{W}),
estimates of E[Y | X = x] (\code{Y.hat}) and E[W | X = x] (\code{W.hat}),
and the localized predictions of the causal forest E[Y_1 - Y_0 | X = x]
(\code{tau.hat})
}
\description{
Estimate heterogeneous treatment effect using Doubly Robust Estimation
(Kennedy 2022) using \code{cv.glmnet} for estimate construction
}
\examples{
# Simulate data
n <- 2000
p <- 20
X <- matrix(rnorm(n * p), n, p)
# CATE varies along one dim only.
tau_ex <- function(x) {
  1 / (1 + exp(-x))
}
TAU <- tau_ex(X[, 3])
# Propensity  and Outcome vary along 2 and 5 dimensions only.
W <- rbinom(n, 1, 1 / (1 + exp(-X[, 1] - X[, 2])))
Y <- pmax(X[, 2] + X[, 3], 0) + rowMeans(X[, 4:6]) / 2 + W * TAU + rnorm(n)

drl <- dr_learner(X, Y, W)

}
\references{
Kennedy, Edward H. (2022) "Towards optimal doubly robust
estimation of heterogeneous causal effects".
\url{https://arxiv.org/abs/2004.14497}.
}
\seealso{
cv.glmnet
}
